+++ TSP_Experiment/README.md
@@ -51,7 +51,7 @@ El éxito del algoritmo ACO depende de encontrar un equilibrio entre explorar nu
 - n_ants debe adaptarse al tamaño del problema.
 
 
-## 📌 Conclusiones y Recomendaciones
+## 📌 Conclusiones y Recomendaciones por Instancia
 
 ### 🧠 BERLIN52
 - 📉 GAP promedio general: 16.19%
@@ -65,16 +65,16 @@ El éxito del algoritmo ACO depende de encontrar un equilibrio entre explorar nu
 **Recomendación:** Usar esta configuración balanceada para `berlin52`.
 
 
-![GAP vs n_ants - berlin52](./grafico_gap_vs_n_ants_berlin52.png)
+![GAP vs n_ants - berlin52](./png/grafico_gap_vs_n_ants_berlin52.png)
 
 
-![GAP vs alpha - berlin52](./grafico_gap_vs_alpha_berlin52.png)
+![GAP vs alpha - berlin52](./png/grafico_gap_vs_alpha_berlin52.png)
 
 
-![GAP vs beta - berlin52](./grafico_gap_vs_beta_berlin52.png)
+![GAP vs beta - berlin52](./png/grafico_gap_vs_beta_berlin52.png)
 
 
-![GAP vs ro - berlin52](./grafico_gap_vs_ro_berlin52.png)
+![GAP vs ro - berlin52](./png/grafico_gap_vs_ro_berlin52.png)
 
 ### 🧠 BURMA14
 - 📉 GAP promedio general: 0.57%
@@ -88,16 +88,16 @@ El éxito del algoritmo ACO depende de encontrar un equilibrio entre explorar nu
 **Recomendación:** Usar esta configuración balanceada para `burma14`.
 
 
-![GAP vs n_ants - burma14](./grafico_gap_vs_n_ants_burma14.png)
+![GAP vs n_ants - burma14](./png/grafico_gap_vs_n_ants_burma14.png)
 
 
-![GAP vs alpha - burma14](./grafico_gap_vs_alpha_burma14.png)
+![GAP vs alpha - burma14](./png/grafico_gap_vs_alpha_burma14.png)
 
 
-![GAP vs beta - burma14](./grafico_gap_vs_beta_burma14.png)
+![GAP vs beta - burma14](./png/grafico_gap_vs_beta_burma14.png)
 
 
-![GAP vs ro - burma14](./grafico_gap_vs_ro_burma14.png)
+![GAP vs ro - burma14](./png/grafico_gap_vs_ro_burma14.png)
 
 ### 🧠 CH150
 - 📉 GAP promedio general: 31.65%
@@ -111,16 +111,16 @@ El éxito del algoritmo ACO depende de encontrar un equilibrio entre explorar nu
 **Recomendación:** Usar esta configuración balanceada para `ch150`.
 
 
-![GAP vs n_ants - ch150](./grafico_gap_vs_n_ants_ch150.png)
+![GAP vs n_ants - ch150](./png/grafico_gap_vs_n_ants_ch150.png)
 
 
-![GAP vs alpha - ch150](./grafico_gap_vs_alpha_ch150.png)
+![GAP vs alpha - ch150](./png/grafico_gap_vs_alpha_ch150.png)
 
 
-![GAP vs beta - ch150](./grafico_gap_vs_beta_ch150.png)
+![GAP vs beta - ch150](./png/grafico_gap_vs_beta_ch150.png)
 
 
-![GAP vs ro - ch150](./grafico_gap_vs_ro_ch150.png)
+![GAP vs ro - ch150](./png/grafico_gap_vs_ro_ch150.png)
 
 ### 🧠 KROA100
 - 📉 GAP promedio general: 27.60%
@@ -134,16 +134,16 @@ El éxito del algoritmo ACO depende de encontrar un equilibrio entre explorar nu
 **Recomendación:** Usar esta configuración balanceada para `kroA100`.
 
 
-![GAP vs n_ants - kroA100](./grafico_gap_vs_n_ants_kroA100.png)
+![GAP vs n_ants - kroA100](./png/grafico_gap_vs_n_ants_kroA100.png)
 
 
-![GAP vs alpha - kroA100](./grafico_gap_vs_alpha_kroA100.png)
+![GAP vs alpha - kroA100](./png/grafico_gap_vs_alpha_kroA100.png)
 
 
-![GAP vs beta - kroA100](./grafico_gap_vs_beta_kroA100.png)
+![GAP vs beta - kroA100](./png/grafico_gap_vs_beta_kroA100.png)
 
 
-![GAP vs ro - kroA100](./grafico_gap_vs_ro_kroA100.png)
+![GAP vs ro - kroA100](./png/grafico_gap_vs_ro_kroA100.png)
 
 ### 🧠 PR439
 - 📉 GAP promedio general: 48.64%
@@ -157,27 +157,16 @@ El éxito del algoritmo ACO depende de encontrar un equilibrio entre explorar nu
 **Recomendación:** Usar esta configuración balanceada para `pr439`.
 
 
-![GAP vs n_ants - pr439](./grafico_gap_vs_n_ants_pr439.png)
+![GAP vs n_ants - pr439](./png/grafico_gap_vs_n_ants_pr439.png)
 
 
-![GAP vs alpha - pr439](./grafico_gap_vs_alpha_pr439.png)
+![GAP vs alpha - pr439](./png/grafico_gap_vs_alpha_pr439.png)
 
 
-![GAP vs beta - pr439](./grafico_gap_vs_beta_pr439.png)
+![GAP vs beta - pr439](./png/grafico_gap_vs_beta_pr439.png)
 
 
-![GAP vs ro - pr439](./grafico_gap_vs_ro_pr439.png)
-
-### 🧠 RESULTADOS_PRUEBA
-- 📉 GAP promedio general: 15.67%
-- 🥇 Mejor GAP alcanzado: 1.25% con parámetros:
-  - n_ants: 30
-  - alpha: 1.0
-  - beta: 2.0
-  - ro: 0.1
-- ⏱️ Tiempo medio de ejecución: 51.39 segundos
-
-**Recomendación:** Usar esta configuración balanceada para `resultados_prueba`.
+![GAP vs ro - pr439](./png/grafico_gap_vs_ro_pr439.png)
 
 ### 🧠 ST70
 - 📉 GAP promedio general: 21.61%
@@ -191,16 +180,16 @@ El éxito del algoritmo ACO depende de encontrar un equilibrio entre explorar nu
 **Recomendación:** Usar esta configuración balanceada para `st70`.
 
 
-![GAP vs n_ants - st70](./grafico_gap_vs_n_ants_st70.png)
+![GAP vs n_ants - st70](./png/grafico_gap_vs_n_ants_st70.png)
 
 
-![GAP vs alpha - st70](./grafico_gap_vs_alpha_st70.png)
+![GAP vs alpha - st70](./png/grafico_gap_vs_alpha_st70.png)
 
 
-![GAP vs beta - st70](./grafico_gap_vs_beta_st70.png)
+![GAP vs beta - st70](./png/grafico_gap_vs_beta_st70.png)
 
 
-![GAP vs ro - st70](./grafico_gap_vs_ro_st70.png)
+![GAP vs ro - st70](./png/grafico_gap_vs_ro_st70.png)
 
 ### 🧠 TSP225
 - 📉 GAP promedio general: 40.09%
@@ -214,16 +203,16 @@ El éxito del algoritmo ACO depende de encontrar un equilibrio entre explorar nu
 **Recomendación:** Usar esta configuración balanceada para `tsp225`.
 
 
-![GAP vs n_ants - tsp225](./grafico_gap_vs_n_ants_tsp225.png)
+![GAP vs n_ants - tsp225](./png/grafico_gap_vs_n_ants_tsp225.png)
 
 
-![GAP vs alpha - tsp225](./grafico_gap_vs_alpha_tsp225.png)
+![GAP vs alpha - tsp225](./png/grafico_gap_vs_alpha_tsp225.png)
 
 
-![GAP vs beta - tsp225](./grafico_gap_vs_beta_tsp225.png)
+![GAP vs beta - tsp225](./png/grafico_gap_vs_beta_tsp225.png)
 
 
-![GAP vs ro - tsp225](./grafico_gap_vs_ro_tsp225.png)
+![GAP vs ro - tsp225](./png/grafico_gap_vs_ro_tsp225.png)
 
 ### 🧠 ULYSSES16
 - 📉 GAP promedio general: 0.55%
@@ -237,16 +226,16 @@ El éxito del algoritmo ACO depende de encontrar un equilibrio entre explorar nu
 **Recomendación:** Usar esta configuración balanceada para `ulysses16`.
 
 
-![GAP vs n_ants - ulysses16](./grafico_gap_vs_n_ants_ulysses16.png)
+![GAP vs n_ants - ulysses16](./png/grafico_gap_vs_n_ants_ulysses16.png)
 
 
-![GAP vs alpha - ulysses16](./grafico_gap_vs_alpha_ulysses16.png)
+![GAP vs alpha - ulysses16](./png/grafico_gap_vs_alpha_ulysses16.png)
 
 
-![GAP vs beta - ulysses16](./grafico_gap_vs_beta_ulysses16.png)
+![GAP vs beta - ulysses16](./png/grafico_gap_vs_beta_ulysses16.png)
 
 
-![GAP vs ro - ulysses16](./grafico_gap_vs_ro_ulysses16.png)
+![GAP vs ro - ulysses16](./png/grafico_gap_vs_ro_ulysses16.png)
 
 ### 🧠 ULYSSES22
 - 📉 GAP promedio general: 2.28%
@@ -260,13 +249,13 @@ El éxito del algoritmo ACO depende de encontrar un equilibrio entre explorar nu
 **Recomendación:** Usar esta configuración balanceada para `ulysses22`.
 
 
-![GAP vs n_ants - ulysses22](./grafico_gap_vs_n_ants_ulysses22.png)
+![GAP vs n_ants - ulysses22](./png/grafico_gap_vs_n_ants_ulysses22.png)
 
 
-![GAP vs alpha - ulysses22](./grafico_gap_vs_alpha_ulysses22.png)
+![GAP vs alpha - ulysses22](./png/grafico_gap_vs_alpha_ulysses22.png)
 
 
-![GAP vs beta - ulysses22](./grafico_gap_vs_beta_ulysses22.png)
+![GAP vs beta - ulysses22](./png/grafico_gap_vs_beta_ulysses22.png)
 
 
-![GAP vs ro - ulysses22](./grafico_gap_vs_ro_ulysses22.png)
+![GAP vs ro - ulysses22](./png/grafico_gap_vs_ro_ulysses22.png)
