+++ TSP_Experiment/png/grafico_gap_vs_alpha_burma14.png
