+++ TSP_Experiment/png/grafico_gap_vs_alpha_ch150.png
