+++ TSP_Experiment/png/grafico_gap_vs_alpha_kroA100.png
