+++ TSP_Experiment/png/grafico_gap_vs_alpha_st70.png
