+++ TSP_Experiment/png/grafico_gap_vs_beta_burma14.png
