+++ TSP_Experiment/png/grafico_gap_vs_beta_ch150.png
