+++ TSP_Experiment/png/grafico_gap_vs_beta_kroA100.png
