+++ TSP_Experiment/png/grafico_gap_vs_beta_st70.png
