+++ TSP_Experiment/png/grafico_gap_vs_beta_ulysses22.png
