+++ TSP_Experiment/png/grafico_gap_vs_ro_ch150.png
