+++ TSP_Experiment/png/grafico_gap_vs_ro_kroA100.png
