+++ TSP_Experiment/png/grafico_gap_vs_ro_st70.png
