+++ TSP_Experiment/png/grafico_gap_vs_ro_ulysses22.png
