--- gridsearch_resultados_prueba.csv
@@ -1,55 +1,0 @@
-archivo,dimension,mejor_distancia,optimo,gap_%,tiempo_segundos,n_ants,alpha,beta,ro,iterations
-berlin52.tsp,52,11437.0,7542.0,51.64,33.752,10,0.5,1.0,0.1,1000
-berlin52.tsp,52,11059.0,7542.0,46.63,82.801,10,0.5,1.0,0.3,1000
-berlin52.tsp,52,10191.0,7542.0,35.12,38.354,10,0.5,1.0,0.5,1000
-berlin52.tsp,52,8742.0,7542.0,15.91,29.373,10,0.5,2.0,0.1,1000
-berlin52.tsp,52,8676.0,7542.0,15.04,29.654,10,0.5,2.0,0.3,1000
-berlin52.tsp,52,8344.0,7542.0,10.63,29.279,10,0.5,2.0,0.5,1000
-berlin52.tsp,52,8099.0,7542.0,7.39,29.549,10,1.0,1.0,0.1,1000
-berlin52.tsp,52,8801.0,7542.0,16.69,29.509,10,1.0,1.0,0.3,1000
-berlin52.tsp,52,8512.0,7542.0,12.86,29.252,10,1.0,1.0,0.5,1000
-berlin52.tsp,52,7676.0,7542.0,1.78,30.242,10,1.0,2.0,0.1,1000
-berlin52.tsp,52,8301.0,7542.0,10.06,29.751,10,1.0,2.0,0.3,1000
-berlin52.tsp,52,7997.0,7542.0,6.03,29.669,10,1.0,2.0,0.5,1000
-berlin52.tsp,52,8339.0,7542.0,10.57,29.892,10,1.5,1.0,0.1,1000
-berlin52.tsp,52,8405.0,7542.0,11.44,30.727,10,1.5,1.0,0.3,1000
-berlin52.tsp,52,9435.0,7542.0,25.1,31.189,10,1.5,1.0,0.5,1000
-berlin52.tsp,52,8228.0,7542.0,9.1,30.82,10,1.5,2.0,0.1,1000
-berlin52.tsp,52,8411.0,7542.0,11.52,31.103,10,1.5,2.0,0.3,1000
-berlin52.tsp,52,8273.0,7542.0,9.69,31.963,10,1.5,2.0,0.5,1000
-berlin52.tsp,52,11840.0,7542.0,56.99,59.09,20,0.5,1.0,0.1,1000
-berlin52.tsp,52,10270.0,7542.0,36.17,59.537,20,0.5,1.0,0.3,1000
-berlin52.tsp,52,10823.0,7542.0,43.5,58.442,20,0.5,1.0,0.5,1000
-berlin52.tsp,52,8908.0,7542.0,18.11,56.065,20,0.5,2.0,0.1,1000
-berlin52.tsp,52,8609.0,7542.0,14.15,56.978,20,0.5,2.0,0.3,1000
-berlin52.tsp,52,8615.0,7542.0,14.23,57.511,20,0.5,2.0,0.5,1000
-berlin52.tsp,52,8319.0,7542.0,10.3,56.033,20,1.0,1.0,0.1,1000
-berlin52.tsp,52,7993.0,7542.0,5.98,57.444,20,1.0,1.0,0.3,1000
-berlin52.tsp,52,8065.0,7542.0,6.93,57.069,20,1.0,1.0,0.5,1000
-berlin52.tsp,52,7701.0,7542.0,2.11,57.725,20,1.0,2.0,0.1,1000
-berlin52.tsp,52,7676.0,7542.0,1.78,55.491,20,1.0,2.0,0.3,1000
-berlin52.tsp,52,8169.0,7542.0,8.31,56.326,20,1.0,2.0,0.5,1000
-berlin52.tsp,52,8373.0,7542.0,11.02,57.398,20,1.5,1.0,0.1,1000
-berlin52.tsp,52,8552.0,7542.0,13.39,57.054,20,1.5,1.0,0.3,1000
-berlin52.tsp,52,9211.0,7542.0,22.13,59.903,20,1.5,1.0,0.5,1000
-berlin52.tsp,52,7745.0,7542.0,2.69,55.618,20,1.5,2.0,0.1,1000
-berlin52.tsp,52,8371.0,7542.0,10.99,56.589,20,1.5,2.0,0.3,1000
-berlin52.tsp,52,7919.0,7542.0,5.0,60.134,20,1.5,2.0,0.5,1000
-berlin52.tsp,52,11517.0,7542.0,52.7,66.826,30,0.5,1.0,0.1,1000
-berlin52.tsp,52,10533.0,7542.0,39.66,60.718,30,0.5,1.0,0.3,1000
-berlin52.tsp,52,11067.0,7542.0,46.74,60.865,30,0.5,1.0,0.5,1000
-berlin52.tsp,52,8303.0,7542.0,10.09,61.081,30,0.5,2.0,0.1,1000
-berlin52.tsp,52,8414.0,7542.0,11.56,61.352,30,0.5,2.0,0.3,1000
-berlin52.tsp,52,8494.0,7542.0,12.62,62.916,30,0.5,2.0,0.5,1000
-berlin52.tsp,52,7957.0,7542.0,5.5,67.236,30,1.0,1.0,0.1,1000
-berlin52.tsp,52,8084.0,7542.0,7.19,64.504,30,1.0,1.0,0.3,1000
-berlin52.tsp,52,8206.0,7542.0,8.8,62.271,30,1.0,1.0,0.5,1000
-berlin52.tsp,52,7636.0,7542.0,1.25,62.266,30,1.0,2.0,0.1,1000
-berlin52.tsp,52,7679.0,7542.0,1.82,61.474,30,1.0,2.0,0.3,1000
-berlin52.tsp,52,7679.0,7542.0,1.82,61.401,30,1.0,2.0,0.5,1000
-berlin52.tsp,52,7714.0,7542.0,2.28,61.11,30,1.5,1.0,0.1,1000
-berlin52.tsp,52,8493.0,7542.0,12.61,61.187,30,1.5,1.0,0.3,1000
-berlin52.tsp,52,8833.0,7542.0,17.12,64.533,30,1.5,1.0,0.5,1000
-berlin52.tsp,52,7830.0,7542.0,3.82,63.617,30,1.5,2.0,0.1,1000
-berlin52.tsp,52,8182.0,7542.0,8.49,64.34,30,1.5,2.0,0.3,1000
-berlin52.tsp,52,8363.0,7542.0,10.89,66.11,30,1.5,2.0,0.5,1000
